requirements.txt
backoff==2.2.1
datasets==3.2.0
faiss_cpu==1.9.0.post1
ir_datasets==0.5.9
json5==0.10.0
nltk==3.9.1
numpy==2.2.5
openai==1.78.1
protobuf==6.31.0
sentence_transformers==3.3.1
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0.dev0
trl==0.13.0
vllm==0.6.4.post2.dev357+g0a56bcc0
